Refactor evolution and outsource mutation algorithms
