+++ python/src/solver/evolution.py
@@ -2,10 +2,12 @@ import logging
 import os
 import shutil
 from random import random
+from typing import Optional
 
 from src.drawing.draw import DrawEmbedding
+from src.embedding.embedding import Embedding
 from src.graph.test_graph import TestGraph
-from src.solver.embedding_solver import EmbeddingSolver
+from src.solver.embedding_solver import EmbeddingSolver, EvolutionParams
 from src.util.logging import init_logger
 
 init_logger()
@@ -14,12 +16,15 @@ logger = logging.getLogger('evolution')
 
 ################################# Params #######################################
 
+params = EvolutionParams(
+    population_size=7,
+    max_mutation_trials=30,
+    mutation_trials_until_extend_to_free_neighbors=15
+)
+
 max_total = 1
-max_mutations_trials = 30
-population_size = 7
 max_generations = 300
 remove_redundant_nodes_probability = 0.1
-mutation_trials_until_extend_to_free_neighbors = int(max_mutations_trials / 2)
 
 
 ############################### Evolution ######################################
@@ -56,7 +61,6 @@ def main(d: DrawEmbedding) -> bool:
 
     solver = EmbeddingSolver(H)
     solver.initialize_embedding()
-    solver.local_maximum()
     save_embedding(*solver.get_embedding(), d, -1,
                    title=f'Initial embedding')
 
@@ -68,72 +72,40 @@ def main(d: DrawEmbedding) -> bool:
 
     # --- Start solver
     for i in range(max_generations):
-        logger.info('')
-        logger.info(f'🔄 Generation: {i}')
+        child = do_one_generation(i, solver)
 
-        # Generate children for one population
-        population = []  # list of Embeddings
-        last_trial_used = False
-        for _ in range(population_size):
-            logger.info('')
-            logger.info(f'--- Try find a new viable mutation')
-            mutation = None
+        if not child:
+            logger.info('🔳 Stopping algorithm...')
+            return False
 
-            for trial, mutation in enumerate(range(max_mutations_trials)):
-                # Do one mutation
-                logger.info('--- MUTATION')
-                mutation = solver.extend_random_supernode()
-                if mutation:
-                    population.append(mutation)
-                    break  # try to construct next child
-                elif trial >= mutation_trials_until_extend_to_free_neighbors:
-                    solver.extend_random_supernode_to_free_neighbors()
-
-            if not mutation:
-                if not len(population):
-                    if not last_trial_used:
-                        # Before all fails: try to remove unnecessary supernode nodes
-                        # and try once more
-                        solver.remove_redundant_supernode_nodes()
-                        logger.info(f'🔳 Last trial, remove redundant nodes')
-                        continue
-                    else:
-                        logger.info(f'🔳 All {max_mutations_trials} mutations failed, '
-                                    'could not construct a single child -> Abort')
-                        return False
-                else:
-                    logger.info(f'🔳 {max_mutations_trials} mutations to construct '
-                                ' a new child failed, will continue '
-                                f'with smaller population: {len(population)}/{population_size}')
-                    # break early since it is improbable that we will be able
-                    # to generate more children
-                    break
-
-        # Choose best child
-        improvements = [mutation.try_embed_missing_edges() for mutation in population]
-        best_mutation_index = improvements.index(max(improvements))
-        best_mutation = population[best_mutation_index]
-
-        # Leave room for next generation
-        if random() < remove_redundant_nodes_probability:
-            logger.info('Try to remove redundant supernode nodes')
-            best_mutation.remove_redundant_supernode_nodes()
-        # best_mutation.remove_unnecessary_edges_between_supernodes()
-
-        solver.commit(best_mutation)
-
-        # Save every x generations
+        solver.commit(child)
         save_embedding(*solver.get_embedding(), d, i,
                        title=f'Generation {i}')
-
-        # Check if we found valid embedding
-        if best_mutation.is_valid_embedding():
+        # Check if done
+        if child.is_valid_embedding():
             logger.info('🎉🎉🎉🎉🎉🎉 Found embedding')
             return True
         else:
             logger.info('✅ Generation passed')
 
     return False
+
+
+def do_one_generation(i: int, solver: EmbeddingSolver) -> Optional[Embedding]:
+    logger.info('')
+    logger.info(f'🔄 Generation: {i}')
+
+    child = solver.generate_population_and_select(params)
+    if not child:
+        return None
+
+    # Leave "room" on graph for next generation
+    if random() < remove_redundant_nodes_probability:
+        logger.info('Try to remove redundant supernode nodes')
+        child.remove_redundant_supernode_nodes()
+        # best_mutation.remove_unnecessary_edges_between_supernodes()
+
+    return child
 
 
 ################################ Output ########################################
